backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0

tensorflow-cpu==2.14.0
opencv-python-headless==4.8.0.76

numpy==1.26.4
python-dotenv
